+++ example/grpc-web/pubspec.yaml
@@ -6,10 +6,9 @@ environment:
   sdk: '>=2.0.0 <3.0.0'
 
 dependencies:
-  grpc:
-    path: ../../
+  grpc: 2.0.0
   protobuf: ^0.13.4
 
 dev_dependencies:
-  build_runner: ^0.10.0
-  build_web_compilers: '^0.4.1'
+  build_runner:
+  build_web_compilers:
