+++ test/src/server_utils.dart
@@ -17,12 +17,10 @@ import 'dart:async';
 
 import 'package:grpc/src/client/http2_connection.dart';
 import 'package:grpc/src/shared/message.dart';
-import 'package:grpc/src/shared/streams.dart';
 import 'package:http2/transport.dart';
 import 'package:test/test.dart';
 
 import 'package:grpc/grpc.dart';
-import 'package:grpc/src/client/transport/http2_transport.dart';
 
 import 'utils.dart';
